Refactor to use custom hook for data logic
